# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [Orders]
      summary: List orders (admin, or filtered by customer phone)
      parameters:
        - name: phone
          in: query
          required: false
          schema:
            type: string
          description: Filter orders by customer phone number.
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createOrder
      tags: [Orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}:
    patch:
      operationId: updateOrderStatus
      tags: [Orders]
      summary: Update an order's status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /orders/{id}/items:
    post:
      operationId: addOrderItems
      tags: [Orders]
      summary: Add items to a recent order within the "نسيت غرض" grace window
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddOrderItemsInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input or grace window elapsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    OrderItemInput:
      type: object
      required: [id, name, price, unit, qty]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: integer
        unit:
          type: string
        qty:
          type: integer
          minimum: 1
    OrderInput:
      type: object
      # customerPhone is intentionally NOT part of the input — the server
      # always derives it from the authenticated customer's bearer token,
      # so a client can never place an order (or earn/spend points) under
      # someone else's phone number.
      required: [items, deliveryType]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
        deliveryType:
          type: string
          enum: [standard, express]
        latitude:
          type: number
        longitude:
          type: number
        redeem:
          type: ["string", "null"]
          enum: [discount, free_delivery, null]
        note:
          type: ["string", "null"]
          maxLength: 500
          description: Optional customer note attached to the order.
        pickupTime:
          type: ["string", "null"]
          maxLength: 100
          description: Customer-chosen delivery/pickup window.
        walletApplied:
          type: integer
          minimum: 0
          description: Amount of in-app wallet credit to spend on this order, in IQD.
        storeId:
          type: ["integer", "null"]
          description: The store this single-store cart belongs to.
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: ["قيد التحضير", "في الطريق", "تم التسليم"]
    Order:
      type: object
      required:
        - id
        - customerPhone
        - items
        - subtotal
        - deliveryFee
        - deliveryType
        - discountApplied
        - pointsEarned
        - pointsRedeemed
        - total
        - status
        - createdAt
      properties:
        id:
          type: integer
        customerPhone:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        subtotal:
          type: integer
        deliveryFee:
          type: integer
        deliveryType:
          type: string
          enum: [standard, express]
        discountApplied:
          type: integer
        pointsEarned:
          type: integer
        pointsRedeemed:
          type: integer
        redemptionType:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
          description: Optional customer note attached to the order.
        total:
          type: integer
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        status:
          type: string
        pickupTime:
          type: ["string", "null"]
        walletApplied:
          type: integer
        storeId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    AddOrderItemsInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"